+++ src/Commands/Tempo/AttributesCommand.php
@@ -3,15 +3,38 @@
 namespace App\Commands\Tempo;
 
 use App\Commands\Command;
+use App\Repositories\TempoRepository;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use App\Services\Validators\Tempo\AttributesValidator;
 
 /**
+ * Class AttributesCommand
+ *
  * @author Mohamed Abdul-Fattah <csmohamed8@gmail.com>
  */
 class AttributesCommand extends Command
 {
+
+    /**
+     * @var AttributesValidator
+     */
+    private $validator;
+
+    /**
+     * @var TempoRepository
+     */
+    private $repository;
+
+    public function __construct(AttributesValidator $validator, TempoRepository $repository)
+    {
+        parent::__construct();
+
+        $this->validator  = $validator;
+        $this->repository = $repository;
+    }
+
     /**
      * Configure attributes command
      */
@@ -19,7 +42,7 @@ class AttributesCommand extends Command
     {
         $this->setName('tempo:attributes')
              ->setDescription('Add attributes to worklog request payload, to be sent with tempo:sync')
-             ->addUsage('{"attributes":"_Role_":{"name":"Role","value":"Developer"}}}')
+             ->addUsage('\'{"attributes":{"_Role_":{"name":"Role","value":"Developer"}}}\'')
              ->addArgument(
                  'attributes',
                  InputArgument::REQUIRED,
@@ -34,6 +57,14 @@ class AttributesCommand extends Command
    */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
+        $attributes = $input->getArgument('attributes');
+
+        $this->validator->isJson($attributes);
+
+        $output->writeln('<comment>Saving attributes...</comment>');
+        $this->repository->saveAttributes($attributes);
+        $output->writeln('<info>Attributes saved successfully</info>');
+
         return self::EXIT_SUCCESS;
     }
 }