+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Dependency injector
+- `tempo:attributes` command to save extra attributes for logging
+
 ## [0.2.2] - 2020-03-11
 ### Fixed
 - Sync command with unauthorized response